Переносит работу обработчиков карточек предложений в offers-list
